# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Restaurant Platform API
  description: |
    Phase 1 backend API — auth, restaurant leads, durable city scraping, human review,
    demo sites, and plain-text-authored outreach sequences.

    This document covers the Phase 1 lead-to-outreach and reservation surface, the
    static-token company consultation API used by the corporate voice agent, and the
    internal consultation-calendar controls. The router remains the source of truth for
    legacy public asset routes outside this workflow specification.

    ## Authentication
    Protected routes use `Authorization: Bearer <access_token>` from signup or login.

    ## Roles
    - `internal_admin` — full restaurant admin (seed via `make seed-admin`)
    - `restaurant_owner` — scoped to assigned restaurants only
    - `developer` — health endpoints (local/test signup only)

    ## Lead ingestion and outreach flow (admin)
    1. `POST /api/v1/scrape-jobs` — create or return the active durable city/niche scrape job.
    2. The scrape worker uses Google Places first and Apollo only as best-effort enrichment
       for missing contact details. Missing credentials or Apollo provider failures never
       discard or block verified Places leads. The worker pauses at 500 combined provider
       requests and resumes after its durable `resume_at`.
    3. Any inferred-business restaurant with a name, valid email, recorded consent evidence,
...
       the demo bearer token.
...

    ## Bulk outreach (one-click)
...

...
servers:
...
paths:
  /api/v1/auth/signup:
    post:
      tags: [Auth]
      summary: Sign up
      description: |
        Self-signup for `restaurant_owner` (all envs) and `developer` (local/test only).
        `internal_admin` cannot be self-assigned — use `make seed-admin`.
      operationId: authSignup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
            example:
              email: owner@local.test
              password: password123
              full_name: Demo Restaurant Owner
              role: restaurant_owner
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Log in
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: admin@local.test
              password: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Current session identity
      operationId: authMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Principal from JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMe'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/admin/me:
    get:
      tags: [Admin]
      summary: Admin profile
      operationId: adminMe
      security:
        - bearerAuth: []
      description: Requires role `internal_admin`.
      responses:
        '200':
          description: Admin user record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/consultation-calendar/{month}:
    parameters:
      - name: month
        in: path
        required: true
        description: Calendar month in the configured consultation timezone.
        schema:
          type: string
          pattern: '^\d{4}-(0[1-9]|1[0-2])$'
          example: '2026-08'
    get:
      tags: [Consultation Calendar]
      summary: Get a monthly consultation calendar
      operationId: getAdminConsultationCalendar
      security:
        - bearerAuth: []
      description: |
        Requires `internal_admin`. Returns every configured candidate in the month,
        persisted availability intent, confirmed-booking state, and any historical
        confirmed booking that falls outside the current candidate grid. Missing
        override rows use the configured business-hours default.
      responses:
        '200':
          description: Monthly consultation calendar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationCalendar'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    put:
      tags: [Consultation Calendar]
      summary: Save monthly consultation availability
      operationId: putAdminConsultationCalendar
      security:
        - bearerAuth: []
      description: |
        Requires `internal_admin`. Uses optimistic concurrency: send the `revision`
        returned by GET as `expected_revision`. The request must contain every future
        on-grid candidate for the month. A canonical same-month candidate that became
        past after GET is accepted and ignored, allowing a save to cross a slot boundary.
        Duplicate, off-grid, or wrong-offset values are rejected. Saving availability
        never cancels or changes a confirmed consultation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsultationCalendarUpdate'
      responses:
        '200':
          description: Saved monthly consultation calendar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationCalendar'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/ConsultationCalendarConflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/company/consultations/availability:
    get:
      tags: [Company Consultations]
      summary: List available company consultation slots
      operationId: getCompanyConsultationAvailability
      security:
        - consultationStaticBearer: []
      parameters:
        - name: date
          in: query
          description: Optional starting local date; defaults to today in the configured timezone.
          schema:
            type: string
            format: date
        - name: days
          in: query
          description: Number of weekdays to scan, capped by the configured availability horizon.
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Available future on-grid slots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationAvailabilityResult'
        '400':
          description: Invalid date or availability request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationStatusError'
        '401':
          description: Missing or incorrect static bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationStatusError'
  /api/v1/company/consultations/availability/check:
    get:
      tags: [Company Consultations]
      summary: Check one company consultation slot
      operationId: checkCompanyConsultationAvailability
      security:
        - consultationStaticBearer: []
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: time
          in: query
          required: true
          description: Local time on the configured consultation grid.
          schema:
            type: string
            example: '09:30'
      responses:
        '200':
          description: Slot state and alternatives when unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationCheckResult'
        '400':
          description: Missing, invalid, off-grid, or past slot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationStatusError'
        '401':
          description: Missing or incorrect static bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationStatusError'
  /api/v1/company/consultations:
    post:
      tags: [Company Consultations]
      summary: Book a company consultation
      operationId: bookCompanyConsultation
      security:
        - consultationStaticBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsultationBookRequest'
      responses:
        '201':
          description: Confirmed consultation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationBookSuccess'
        '400':
          description: Invalid booking request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationStatusError'
        '401':
          description: Missing or incorrect static bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationStatusError'
        '409':
          description: Slot is disabled or overlaps an existing confirmed consultation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationBookConflict'
  /api/v1/consultations/availability:
    get:
      deprecated: true
      tags: [Company Consultations]
      summary: List consultation slots (legacy alias)
      operationId: getLegacyConsultationAvailability
      description: Deprecated alias of `/api/v1/company/consultations/availability`.
      security:
        - consultationStaticBearer: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: days
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Available future on-grid slots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationAvailabilityResult'
        '400':
          description: Invalid availability request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationStatusError'
        '401':
          description: Missing or incorrect static bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationStatusError'
  /api/v1/consultations/availability/check:
    get:
      deprecated: true
      tags: [Company Consultations]
      summary: Check a consultation slot (legacy alias)
      operationId: checkLegacyConsultationAvailability
      description: Deprecated alias of `/api/v1/company/consultations/availability/check`.
      security:
        - consultationStaticBearer: []
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: time
          in: query
          required: true
          schema:
            type: string
            example: '09:30'
      responses:
        '200':
          description: Slot state and alternatives when unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationCheckResult'
        '400':
          description: Invalid slot request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationStatusError'
        '401':
          description: Missing or incorrect static bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationStatusError'
...
  /api/v1/restaurants/{id}/status:
    parameters:
      - $ref: '#/components/parameters/restaurantId'
    patch:
      tags: [Restaurants]
      summary: Update restaurant lifecycle status
      operationId: updateRestaurantStatus
      security:
        - bearerAuth: []
      description: Requires role `internal_admin`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRestaurantStatusRequest'
            example:
              status: demo_ready
      responses:
        '200':
          description: Updated restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from signup or login
    consultationStaticBearer:
      type: http
      scheme: bearer
      description: Static token configured as `TUVI_API_TOKEN`; used by approved company consultation clients such as the corporate voice agent.
...